% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi_opt.R
\name{tpi_opt}
\alias{tpi_opt}
\title{Optimizes the radius of a TPI used to predict another variable}
\usage{
tpi_opt(
  min_range,
  max_range,
  dem,
  raster,
  n_sample,
  n_bayesianiterations,
  kappa_bayesian,
  correlation_coefficient = "spearman"
)
}
\arguments{
\item{min_range}{Numeric. The minimum tpi-radius.}

\item{max_range}{Numeric. The maximum tpi-radius. In between those to values, the optimal will be found.}

\item{dem}{terra::SpatRaster. The Digital Elevation Model used to calculate the TPI.}

\item{raster}{terra::SpatRaster. The target raster to correlate against. E.g. soil moisture, NDVI, erosion rates, snowdepth...}

\item{n_sample}{Numeric. Number of pixels for which the TPI get calculated at each step. Increasing this increases the computation time and accuracy}

\item{n_bayesianiterations}{Numeric. Number of iterations for the bayesian optimizer. Increasing this increases the computation time and accuracy}

\item{kappa_bayesian}{Numeric. The kappa parameter for the acquisition function.}

\item{correlation_coefficient}{String. The type of  correlation to use. (spearman, pearson. rmse_linear, r2_linear, rmse_quad, r2_quad, rmse_cubic, r2_cubic)}
}
\value{
A list object of class \code{BayesianOptimization} containing the optimization results.
}
\description{
Uses Bayesian optimization and sampling of single pixels at certain using different radii.
}
