% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi_sample.R
\name{tpi_sample}
\alias{tpi_sample}
\title{optimizes the radius of a TPI used to to predict another variable}
\usage{
tpi_sample(
  raster_dsm,
  raster_target,
  sample_num,
  radius,
  relationship = "spearman"
)
}
\arguments{
\item{raster_dsm}{terra::SpatRaster. The Digital Elevation Model that will be used to compute TPI. Both rasters (raster_dsm, raster_target) should have the same grid, resolution, crs and unit.}

\item{raster_target}{terra::SpatRaster. The target raster to correlate against, e.g. NDVI, soil moisture, snow depth, If the target variable is only available as a SpatVect, please rasterize. Both rasters (raster_dsm, raster_target) should have the same grid, resolution, crs and unit.}

\item{sample_num}{Numeric. Number of pixels for which the TPI get calculated. Increasing n_sample increases computational effort and accuracy.}

\item{radius}{Numeric. The radius used to compute TPI}

\item{relationship}{String. The type of correlation to use (e.g., "pearson", "spearman", "rmse_linear", "r2_linear", "rmse_quad", "r2_quad", "rmse_cubic", "r2_cubic").}
}
\description{
Uses Bayesian optimization and sampling of single pixels at certain using different radii
}
