% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi_opt.R
\name{tpi_opt}
\alias{tpi_opt}
\title{optimizes the radius of a TPI used to to predict another variable}
\usage{
tpi_opt(
  min_range,
  max_range,
  dem,
  raster,
  n_sample,
  n_bayesianiterations,
  kappa_bayesian,
  correlation_coefficient = "spearman"
)
}
\arguments{
\item{min_range}{Numeric. The minimum tpi - radius for optimization. The function will only find the optimal radius above this value.}

\item{max_range}{Numeric. The maximum tpi-radius for optimization.The function will only find the optimal radius smaller than this value.}

\item{dem}{terra::SpatRaster. The Digital Elevation Model that will be used to compute TPI}

\item{raster}{terra::SpatRaster. The target raster to correlate against, e.g. NDVI, soil moisture, snow depth, If the target variable is only available as a SpatVect, please rasterize.}

\item{n_sample}{Numeric. Number of pixels for which the TPI get calculated at each step. Increasing n_sample increases computational effort and accuracy}

\item{n_bayesianiterations}{Numeric. Number of iterations for the optimizer, increasing n_bayesianiterations increases computational effort and accuracy}

\item{kappa_bayesian}{Numeric. The kappa parameter for the acquisition function. Higher kappa makes the optimizer more explorativ and less likely to miss out a maxima.}

\item{correlation_coefficient}{String. The type of correlation to use (e.g., "pearson", "spearman", "rsme_linear", "r2_linear", "rsme_quad", "r2_quad", "rmse_cubic", "r2_cubic").}
}
\description{
Uses Bayesian optimization and sampling of single pixels at certain using different radii
}
