% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi_opt.R
\name{tpi_opt}
\alias{tpi_opt}
\title{Optimizes the radius of a TPI used to predict another variable}
\usage{
tpi_opt(
  min_range,
  max_range,
  dem,
  raster,
  n_sample,
  n_bayesianiterations,
  kappa_bayesian,
  correlation_coefficient = "spearman"
)
}
\arguments{
\item{min_range}{Numeric. The minimum tpi-radius for optimization...}

\item{max_range}{Numeric. The maximum tpi-radius for optimization. The optimal radius between these to values will be found.}

\item{dem}{terra::SpatRaster. The Digital Elevation Model used to calculate the TPI. Both of the used rasters should have the same crs, resolution and grid.}

\item{raster}{terra::SpatRaster. The target raster to correlate against. Eg. erosion rates, soil moisture, snow depth, NDVI}

\item{n_sample}{Numeric. Number of pixels for which the TPI get calculated in each try. Increasing this increases accuracy and computational time.}

\item{n_bayesianiterations}{Numeric. Number of iterations for the bayesian optimizer.}

\item{kappa_bayesian}{Numeric. The kappa parameter for the acquisition function. A higher kappa lead to a more explorative bayesian optimization.}

\item{correlation_coefficient}{String. The type of correlation to use. Options are: pearson, spearman, rmse_linear, r2_linear, rmse_quad, r2_quad, rmse_cubic, r2_cubic}
}
\value{
A list object of class \code{BayesianOptimization} containing the optimization results.
}
\description{
Uses Bayesian optimization and sampling of single pixels at certain using different radii.
}
