% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi_opt.R
\name{tpi_opt}
\alias{tpi_opt}
\title{Optimizes the radius of a TPI used to predict another variable}
\usage{
tpi_opt(
  min_range,
  max_range,
  dem,
  raster,
  n_sample,
  n_bayesianiterations,
  kappa_bayesian,
  correlation_coefficient = "spearman"
)
}
\arguments{
\item{min_range}{Numeric. The minimum tpi-radius for optimization...}

\item{max_range}{Numeric. The maximum tpi-radius for optimization...}

\item{dem}{terra::SpatRaster. The Digital Elevation Model...}

\item{raster}{terra::SpatRaster. The target raster to correlate against...}

\item{n_sample}{Numeric. Number of pixels for which the TPI get calculated...}

\item{n_bayesianiterations}{Numeric. Number of iterations for the optimizer...}

\item{kappa_bayesian}{Numeric. The kappa parameter for the acquisition function...}

\item{correlation_coefficient}{String. The type of correlation to use...}
}
\value{
A list object of class \code{BayesianOptimization} containing the optimization results.
}
\description{
Uses Bayesian optimization and sampling of single pixels at certain using different radii.
}
